requirements.txt
google-generativeai>=0.8.0
openai>=1.0.0
pypdf>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0